lldp: T7194: disable individual interface has no effect if 'all' is defined
Enabling LLDP on all interface except one or more individual interfaces are set
to disable has no effect.

To reproduce use:
  set service lldp interface all
  set service lldp interface eth0 disable

One might assume that LLDP is used in rx/tx mode on all interface but not on
eth0. This is not true as LLDP is not explicitly disabled in the configuration,
it rather is "just skipped" and not configured at all.