interface: T7379: do not request SLAAC default route when only DHCPv6 is set
When an interface runs in DHCPv6 only mode, there is no reason to have a
default installed that was received via SLAAC. If SLAAC is needed, it should
be turned on explicitly.

This bug was only triggered during system boot where a DHCPv6 client address
and a default route to a link-local address was shown in the system. If DHCPv6
was enabled only on an interface while VyOS was already running - no default
route got installed.

(cherry picked from commit de44c6aef249b5c3350a5114a38eee3a761f7de0)