ipsec: T7593: Add dynamic prefix for local and remote traffic selectors
In case when there is no local/remote prefix configured in a tunnel settings,
a protocol configured for such tunnel is ignored.

The correct way to generate the configuration is to set the prefix
to `dynamic` if it was not set. The correct config for the described case is:

```
local_ts = dynamic[gre/]
remote_ts = dynamic[gre/]
```

(cherry picked from commit cdb97173c8ba251c577e30bb324555a418c32828)