ssh: T7839: fix warning on deprecated algorithms during commit
The list calculation of in-use but deprecated SSH hostkey algorithms was wrong.
This was implemented in commit 6deda171e ("ssh: T7839: add deprecation warning
for DSA hostkey-algorithm usage"). It always returned the content of the list
of deprecated algorithms, but not the list of deprecated algorithms actually -
in use - by the configuration. This has been corrected.

Before:
  DEPRECATION WARNING: Support for SSH-DSA keys is deprecated and will
  be removed in VyOS 1.6. Please update affected keys to a supported
  algorithm (e.g., RSA, ECDSA or ED25519) to avoid authentication
  failures after the upgrade. The following hostkey-algorithms are in
  use: ssh-dss, ssh-dss-cert-v01@openssh.com

After:
  DEPRECATION WARNING: Support for SSH-DSA keys is deprecated and will
  be removed in VyOS 1.6. Please update affected keys to a supported
  algorithm (e.g., RSA, ECDSA or ED25519) to avoid authentication
  failures after the upgrade. The following hostkey-algorithms are in
  use: ssh-dss

The generation of the MOTD was not affected!

(cherry picked from commit 7aa35eff59611cb90f2409831650ac39506163aa)