boot-config-loader: T7889: Inform user during login of config-load issues
As T7885 turns out to be a config load related bug the user will only be
informed when entering conf mode that something is off.

  vyos@vyos:~$ configure
  WARNING: There was a config error on boot: saving the configuration now could overwrite data.
  You may want to check and reload the boot config

More information is displayed on tty0 of the router - but not everyone has
access to tty0. This change is about copying the message displayed on tty0 to
the MOTD system.

(cherry picked from commit defc29ef4d42224fbea591fc2909d83e7904c20e)