ssh: T8098: defining cipher rijndael-cbc@lysator.liu.se will break SSH daemon
rijndael-cbc@lysator.liu.se is an alias for the cipher better known as AES
(Advanced Encryption Standard), as implemented before it was named aes256-cbc
by the IETF.

It can be worth mentioning that it has also been deprecated in OpenSSH since
2014 with the release of OpenSSH 6.7 - unfortunately we carried over the lagecy.

Any user who will set "rijndael-cbc@lysator.liu.se" on the CLI will render the
system unaccessible, as sshd will no longer start. To keep the broken CLI
compatible a deprecation warning is shown to the user, and internally we will
move to the supported aes256-cbc cipher. CLI nodes will be removed in VyOS 1.5.

    vyos@vyos# set service ssh ciphers rijndael-cbc@lysator.liu.se
    vyos@vyos# commit

    DEPRECATION WARNING: Support for rijndael-cbc@lysator.liu.se (a pre-
    standard name for aes256-cbc) will be removed in VyOS 1.5; internal
    usage has moved to aes256-cbc.